Add files via upload
I have created a dashboard in Tableau summarizing the road accidents in Victoria. The dashboard provides insights into the total number of accidents and the number of people killed or injured across different speed zones. It also examines the light conditions at the time of the accidents, identifying which condition had the highest number of accidents. Similarly, I have included data on deaths related to various conditions such as postcode and so on.  I have attached the file here you can give your views by observing the file too.